chore: Tidy up pyproject.toml
This removes duplicated values between `[project]` and `[tool.poetry]`,
favouring keeping declarations in `[project]`, since most fields that can
go in either are deprecated in `[tool.poetry]`.

An exception to this is `classifiers` which is moved into
`[tool.poetry]`, to utilise Poetry’s automatic setting of Python and
license classifiers. (This also means that license and most Python
classifiers were removed.)