docs: Make project REUSE v3.3 compliant (#337)
This adds a `REUSE.toml` to specify the license for all files that are
not saying otherwise, moves the LICENSE into `LICENSES/` (and replaces
it with a symlink).

The license section of the `README` has a paragraph added about
REUSE compliance with a link to https://reuse.software/spec-3.3/
for more details.

Finally, a GitHub CI compliance check and a pre-commit hook are added,
both taken from https://reuse.software/dev/
with the pre-commit hook changed to only check altered files, as per
https://github.com/fsfe/reuse-tool/blob/17cd41664794c4c74ade18f9f287114c63e82b7f/README.md#run-as-pre-commit-hook

Fixes https://github.com/openfoodfacts/openfoodfacts-python/issues/185